requirements.txt
python-telegram-bot==20.9
requests==2.31.0
python-dotenv==1.0.1
SQLAlchemy==2.0.29
aiosqlite==0.19.0
tmdbv3api==1.9.0
beautifulsoup4==4.12.3
lxml==5.2.0